third_party/WebKit/LayoutTests/shapedetection/detectshape-HTMLImageElement.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/mock-shapedetection.js"></script>
<body>
<img id='img' src='../media/content/greenbox.png'/>
</body>
<script>

var createTestForImageElement = function(detectorName, detectionResultTest) {
  async_test(function(t) {
    var img = document.getElementById("img");

    var theMock = null;
    mockShapeDetectionReady
      .then(mock => {
        theMock = mock;
        var detector = eval("new " + detectorName + "();");
        return detector;
      })
      .catch(error => {
        assert_unreached("Error creating MockShapeDetection: " + error);
      })
      .then(detector => {
        return detector.detect(img);
      })
      .then(detectionResult => {
        detectionResultTest(detectionResult, theMock);
        t.done();
      })
      .catch(error => {
        assert_unreached("Error during detect(img): " + error);
      });

  }, "Detector detect(HTMLImageElement)");
};

function FaceDetectorDetectionResultTest(detectionResult, mock) {
  const imageReceivedByMock = mock.getFrameData();
  assert_equals(imageReceivedByMock.byteLength, 2500, "Image length");
  const GREEN_PIXEL = 0xFF00FF00;
  assert_equals(imageReceivedByMock[0], GREEN_PIXEL, "Pixel color");
  assert_equals(detectionResult.length, 3, "Number of faces");
}

function BarcodeDetectorDetectionResultTest(detectionResult, mock) {
  assert_equals(detectionResult.length, 2, "Number of barcodes");
  assert_equals(detectionResult[0].rawValue, "cats", "barcode 1");
  assert_equals(detectionResult[1].rawValue, "dogs", "barcode 2");
}

// These tests verify that a Detector's detect() works on an HTMLImageElement.
// Use the mock mojo server implemented in mock-shapedetection.js.
generate_tests(createTestForImageElement, [
  [ "Face", "FaceDetector", FaceDetectorDetectionResultTest ],
  [ "Barcode", "BarcodeDetector", BarcodeDetectorDetectionResultTest ]
]);

</script>
